+++ HISTORY.rst
@@ -11,7 +11,7 @@ History
 * add ``Accept`` header type checking before requests and return ``HTTPNotAcceptable [406]`` if invalid
 * code formatting changes for consistency and cleanup of redundant/misguiding names (#162)
 * add option ``MAGPIE_UI_ENABLED`` allowing to completely disable all ``/ui`` route (enabled by default)
-* swagger response improvements (#126)
+* fix swagger responses status code and description and fix erroneous body (#126)
 * fix invalid member count value returned on ``/groups/{id}`` request
 * fix invalid ``DELETE /users/{usr}/services/{svc}/permissions/{perm}`` request not working
 * add more unittests (#74)